(PC-29516)[PRO] refactor: remove delete priceCategory adapter
